Made InputEvent use floating point coordinates.

+++ core/os/input_event.h
@@ -205,8 +205,8 @@ struct InputEventMouse  {
 
 	InputModifierState mod;
 	int button_mask;
-	int x,y;
-	int global_x,global_y;
+	float x,y;
+	float global_x,global_y;
 	int pointer_index;
 };
 
@@ -221,7 +221,7 @@ struct InputEventMouseButton : public InputEventMouse {
 
 struct InputEventMouseMotion : public InputEventMouse {
 
-	int relative_x,relative_y;
+	float relative_x,relative_y;
 	float speed_x,speed_y;
 };
 
@@ -241,14 +241,14 @@ struct InputEventJoypadButton  {
 struct InputEventScreenTouch  {
 
 	int index;
-	int x,y;
+	float x,y;
 	bool pressed;
 };
 struct InputEventScreenDrag  {
 
 	int index;
-	int x,y;
-	int relative_x,relative_y;
+	float x,y;
+	float relative_x,relative_y;
 	float speed_x,speed_y;
 };
